+++ modules/xbup-parser-tree/src/main/java/org/xbup/lib/parser_tree/XBTreeNode.java
@@ -201,7 +201,7 @@ public class XBTreeNode implements XBEditableBlock, TreeNode, UBStreamable {
                 // Data Block
                 dataMode = XBBlockDataMode.DATA_BLOCK;
                 data = new XBData();
-                data.loadFromStream(stream, dataPartSize.getInt());
+                data.loadFromStream(stream, dataPartSize.getLong());
                 size += dataPartSize.getInt();
             } else if (attrPartSize.getInt() < dataPartSize.getSizeUB()) {
                 throw new XBParseException("Attribute overreached", XBProcessingExceptionType.ATTRIBUTE_OVERFLOW);
